Organization & working on bot player
