+++ src/commands/game_c4/bot_player.rs
@@ -1,0 +1,7 @@
+use super::board::Board;
+
+pub trait BotPlayer {
+    type Token;
+    fn accept(&mut self, board: &Board<Self::Token>);
+    fn decide(&self) -> i32;
+}
