+++ src/commands/game_c4/direction.rs
@@ -1,0 +1,11 @@
+#[derive(Clone, Copy, Debug, PartialEq)]
+pub enum Direction {
+    North,
+    NorthEast,
+    East,
+    SouthEast,
+    South,
+    SouthWest,
+    West,
+    NorthWest,
+}
