+++ src/commands/game_c4/game_status.rs
@@ -1,0 +1,8 @@
+use super::Player;
+
+#[derive(Clone, Copy, Debug, PartialEq)]
+pub enum GameStatus {
+    Closed,
+    Playing,
+    Won { player: Player },
+}
