+++ src/commands/game_c4/mod.rs
@@ -1,7 +1,19 @@
-pub use c4::ConnectFour;
+use board::Board;
+use c4::ConnectFour;
+use direction::Direction;
 pub use discord_hooks::ConnectFourDiscord;
+use discord_message::ConnectFourMessage;
+use game_status::GameStatus;
+use player::Player;
+use token::Token;
 
 mod board;
+mod bot_player;
+mod bot_player_auto;
 mod c4;
+mod direction;
 mod discord_hooks;
 mod discord_message;
+mod game_status;
+mod player;
+mod token;
